requirements.txt
requests
psutil
mcp
aiohttp
fastapi 
uvicorn 
tenacity 
aiolimiter
aiofiles
pydantic
python-dotenv
